gcs_functions/requirements.txt
functions-framework>=3.5.0
google-cloud-storage>=2.16.0
google-cloud-bigquery>=3.17.0
google-cloud-firestore>=2.13.0

